change home page to index and add top viewing
